Removed code warnings Updated RfmUsb version

+++ RfmOta/IOtaService.cs
@@ -22,7 +22,7 @@
 * SOFTWARE.
 */
 
-using RfmUsb;
+using RfmUsb.Net;
 using System;
 using System.IO;
 
@@ -40,8 +40,8 @@ namespace RfmOta
         /// <param name="stream">The stream for the hex file to upload to the device</param>
         /// <param name="crc">The crc calculated for the uploaded flash image</param>
         /// <returns>true if the update succeeds</returns>
-        /// <remarks>The <see cref="IOtaService"/> depends on an <see cref="IRfmUsb"/> instance to transmit and receive packets.
-        /// The containing application must open and close the <see cref="IRfmUsb"/> instance.</remarks>
-        bool OtaUpdate(int outputPower, Stream stream, out uint crc);
+        /// <remarks>The <see cref="IOtaService"/> depends on an <see cref="IRfm69"/> instance to transmit and receive packets.
+        /// The containing application must open and close the <see cref="IRfm69"/> instance.</remarks>
+        bool OtaUpdate(sbyte outputPower, Stream stream, out uint crc);
     }
 }
